+++ modelcheck/src/main/kotlin/de/itemis/mps/gradle/modelcheck/ModelCheckArgs.kt
@@ -8,9 +8,11 @@ import kotlin.test.fail
 
 class ModelCheckArgs(parser: ArgParser) : Args(parser) {
     val models by parser.adding("--model",
-        help = "list of models to check")
+        help = "list of models to check (regexes)")
     val modules by parser.adding("--module",
-        help = "list of modules to check")
+        help = "list of modules to check (regexes)")
+    val excludeModels by parser.adding("--exclude-model", help = "list of models to exclude from check (regexes)")
+    val excludeModules by parser.adding("--exclude-module", help = "list of modules to exclude from check (regexes)")
     val warningAsError by parser.flagging("--warning-as-error", help = "treat model checker warning as errors")
     val dontFailOnError by parser.flagging("--error-no-fail", help = "report errors but don't fail the build")
     val xmlFile by parser.storing("--result-file", help = "stores the result as an JUnit xml file").default<String?>(null)